+++ Wet4/malloc_4.cpp
@@ -504,9 +504,18 @@ void *smalloc(size_t size)
             wilderness_is_free = wilderness_chunk->is_free;
             if (wilderness_is_free)
             {
-                size_for_sbrk = size - wilderness_chunk->block_size;
-                _deleteFromHistogram(wilderness_chunk);
-                //_deleteFromHistogram(wilderness_chunk,false);
+                if (size > wilderness_chunk->block_size)
+                {
+                    size_for_sbrk = size - wilderness_chunk->block_size;
+                    _deleteFromHistogram(wilderness_chunk);
+                }
+                else
+                { // if wilderness_chunk  > size to allocate
+                    MallocMetadata * temp_old_wilderness = wilderness_chunk; // save the start of the current block because wilderness will change
+                    _trySplitBlock(wilderness_chunk, size);
+                    DO_IF_DEBUG(_printDebugInfo(__FUNCTION__, size););
+                    return _metadataToPtr(temp_old_wilderness);
+                }
             }
         }
